refactor: chartModal 경로 변경
